Address round-2 review on #9/#10
Fixes a real 500: limitParam validated with Atoi into an int, so a limit
of 2147483648 passed the >0 check and then wrapped negative through
int32(), reaching Postgres as LIMIT -2147483648. It now parses at the
int32 width the sqlc params take, so out-of-range is a clean 400.
Pre-existing on /v1/posts, now covered on both endpoints.

Also: narrow CLAUDE.md's limit claim to the two endpoints that take one,
have seedFollowedPost return the feed name (search tests re-derived
"alice's Feed" as a literal), and drop the getBookmarks middle man.

+++ internal/api/posts.go
@@ -56,7 +56,7 @@ func handleListPosts(db *database.Queries) authedHandler {
 		}
 		rows, err := db.GetPostsForUser(r.Context(), database.GetPostsForUserParams{
 			UserID: user.ID,
-			Limit:  int32(limit),
+			Limit:  limit,
 			Offset: 0,
 		})
 		if err != nil {
